pyproject.toml
[project]
name = "astrology"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.5",
    "fastapi>=0.116.2",
    "flask>=3.1.2",
    "gdown>=5.2.1",
    "geopy>=2.4.1",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.0",
    "langchain-classic>=1.0.3",
    "langchain-community>=0.3.0",
    "langchain-google-genai>=4.2.1",
    "langchain-qdrant>=1.1.0",
    "pymupdf>=1.27.2",
    "pypdf>=6.9.0",
    "pyswisseph>=2.10.3.2",
    "python-dotenv>=1.1.1",
    "pytz>=2025.2",
    "qdrant-client>=1.17.1",
    "rapidocr-onnxruntime>=1.4.4",
    "swisseph>=0.0.0.dev1",
    "timezonefinder>=8.0.0",
    "uvicorn>=0.35.0",
]
